add safe_stoi to utils/string_utils.h
